tasks/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
bidict==0.22.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
Flask-SocketIO==5.3.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.36
pygame==2.1.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7
redis==4.4.1
six==1.16.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
